refactor: DTO를 Domain으로 이동시킴
