+++ src/main/java/gift/product/presentation/ProductManageController.java
@@ -1,6 +1,7 @@
 package gift.product.presentation;
 
 import gift.product.application.ProductService;
+import gift.product.domain.CreateProductRequestDTO;
 import gift.product.domain.Product;
 import gift.util.CommonResponse;
 import jakarta.validation.Valid;
@@ -48,29 +49,4 @@ public class ProductManageController {
     }
 
 
-    public class CreateProductRequestDTO {
-
-        private static final int MAX_INPUT_LENGTH = 255;
-
-        @NotBlank(message = "이름은 필수 입력 값입니다")
-        private String name;
-
-        @NotNull(message = "가격은 필수 입력 값입니다.")
-        private Double price;
-
-        @Size(max = MAX_INPUT_LENGTH, message = "이미지 URL은 255자를 넘을 수 없습니다.")
-        private String imageUrl;
-
-        public String getName() {
-            return name;
-        }
-
-        public Double getPrice() {
-            return price;
-        }
-
-        public String getImageUrl() {
-            return imageUrl;
-        }
-    }
 }
